Fix missing Content-Type for woff/woff2
Our font assets in S3 have the incorrect media type of
binary/octet-stream, instead of font/woff. This is because the aws s3
cli is unable to detect this content type (see:
https://github.com/aws/aws-cli/issues/1249). Therefore this program
recieves the wrong media type for our fonts and fails to download them.
This modifies the behaviour to fall back to allowing downloads for
response that already have an extension.