Fix attempts to parse xlsx as xml
This is a hacky work around colly's handleOnXML behaviour which considers
any response body with content-type containing the substring "xml" to be
parsed as XML. This is an incorrect assumption for docx, xlsx files
which aren't strictly xml structured and cause parsing errors. Here we
remove "xml" from those content types to prevent them being parsed.