Make the crawler keep within 100 req/s by forcing a delay after each request
The crawler has a number of threads T, across which the 100 requests are
spread. To ensure we don't go over the limit, we need to make sure each threads
waits long enough.

For the sake of simple maths, let's imagine we're aiming for 10 requests per
second. If we have one thread, then we want to send a request every 100ms. That
100ms is made up of an amount of time spent processing and an amount spent
waiting. Analysis of the logs shows a conservative estimate for the time spent
processing is 4ms, therefore the waiting time is 96ms.

If we have two threads, and each thread waited 96ms we'd end up with two
requests every 100ms and 20 reqs. Instead if we double the wait time, we end up
with two requests every 196ms and ~10 req/s. This is ever so slightly over the
target, but in practice the github.com/gocolly/colly/v2 package adds a small
random delay on top of the given delay, which will reduce the request rate
slightly further.

We can generalise this to multiplying the wait time for a single thread by the
number of threads to get close to the target.

For request rate R and threads T the equations becomes

    delay = ((1000/R)-4) * T