Change "url" log field to "crawled_url"
The presence of the "url" field in the log lines was causing its logs not to be
ingested by OpenSearch, because OpenSearch has, for some reason, decided that
when it sees a "url" field it should be expecting an object. This is not the
case for the crawler's log lines.

We expect that other applications DO emit a "url" field with an object and that
OpenSearch's configuration is probably correct. We don't want to change it, so
we make a small tweak to the output of the crawler.