Remove redundant error handling
Error handling removed in htmlHandler, sitemapXmlHandler and scrapeHandler

The error handling will be taken care of by the errorHandler.

When running in synchronous mode the Visit function will return the error which means that it will be processed by the errorHandler and by the function that called Visit.
In async mode the error returned is always nil which is why there is a difference between the error metrics if the error handling in these functions are not removed.