Set the govuk_mirror_last_updated_time after the crawl job has finished
Rather than fetching the last_updated time from the MirrorFreshnessUrl just set the metric when the crawl job is complete.

Remove MIRROR_FRESHNESS_URL and func `fetchMirrorFreshnessMetric`

As the `govuk_mirror_last_updated_time` metric will be pushed after the crawler has completed there is no need for this envvar, so remove it to limit confusion.