Check ResponseMetrics prefix are correct and update TestUpdateMetrics
In order to check that the ReponseMetrics have the correct prefixes the metric needs to have a metric set as it is a GaugeVec. Without a metric set no metric is emitted.
Additionally code has been added to ensure that all the metrics are being emitted as the test would have passed without this check.

The TestUpdateMetrics func has been updated to run the UpdateMirrorResponseStatusCode synchronously rather than async as this will do away with the time.sleep which was in place to prevent a race condition thus simplifying the test code.